Pull Request Refactoring
