+++ common/messages/trade.ts
@@ -1,0 +1,8 @@
+import { TetrominoType } from "../TetrominoType";
+export interface ToClientEvents {
+    sendTradePiece: (tetrominoType: TetrominoType) => void;
+}
+
+export interface ToServerEvents {
+    sendTradePiece: (tetrominoType: TetrominoType) => void;
+}
