+++ common/shared.ts
@@ -3,3 +3,4 @@ export const BOARD_SIZE = 40;
 export const TILE_SIZE = 20;
 
 export type ColoredScore = { color: string; hex: number; points: number };
+
