+++ server/package.json
@@ -17,6 +17,7 @@
 		"dotenv": "^14.3.2",
 		"express": "^4.17.2",
 		"path": "^0.12.7",
+		"prettier": "^2.5.1",
 		"socket.io": "^4.4.1",
 		"typescript": "^4.5.5"
 	},
